+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# 0.19.1 (TBA)
+# 0.19.1 (Apr 24, 2019)
 
 Note: bumped minor version number 19, that indicated the year 2019. The minor versions are the release numbers then (1 for the second release this year).